requirements.txt
asgiref==3.8.1
decorator==5.1.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.1